# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InstaClone API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor de API Principal"
        }
    ],
    "paths": {
...
        "/posts": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Criar novo post",
                "description": "Cria um novo post com imagem e caption opcional. Requer autenticação.",
                "operationId": "21a6ac5f9280a3cf958c3d066b5600ee",
                "requestBody": {
                    "description": "Dados do post a ser criado",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "Imagem do post (máx 5MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "caption": {
                                        "type": "string",
                                        "example": "Que dia lindo! ☀️ #natureza #viagem",
                                        "maxLength": 2200
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "caption": {
                                            "type": "string",
                                            "example": "Que dia lindo! ☀️ #natureza #viagem"
                                        },
                                        "image_url": {
                                            "type": "string",
                                            "example": "/storage/posts/image_1.jpg"
                                        },
                                        "likes_count": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "comments_count": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-05-22T10:30:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-05-22T10:30:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Arquivo inválido ou muito grande"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/posts/{postId}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Obter detalhes do post",
                "description": "Retorna os detalhes completos de um post específico incluindo informações do autor.",
                "operationId": "04e28d03abbc1c8cbe8d01b92b4011b2",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID do post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do post retornado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "caption": {
                                            "type": "string",
                                            "example": "Que dia lindo! ☀️ #natureza #viagem"
                                        },
                                        "image_url": {
                                            "type": "string",
                                            "example": "/storage/posts/image_1.jpg"
                                        },
                                        "likes_count": {
                                            "type": "integer",
                                            "example": 42
                                        },
                                        "comments_count": {
                                            "type": "integer",
                                            "example": 8
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-05-22T10:30:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-05-22T10:30:00Z"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "João Silva"
                                                },
                                                "username": {
                                                    "type": "string",
                                                    "example": "joao_silva"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "joao@example.com"
                                                },
                                                "avatar": {
                                                    "type": "string",
                                                    "example": "/storage/avatars/joao.jpg"
                                                },
                                                "bio": {
                                                    "type": "string",
                                                    "example": "Fotógrafo e viajante"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Post"
                ],
                "summary": "Atualizar post",
                "description": "Atualiza a caption de um post existente. Apenas o autor pode atualizar. Requer autenticação.",
                "operationId": "26847e8ad19915583bd05e580ec8c6b0",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID do post a atualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados a atualizar no post",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "caption"
                                ],
                                "properties": {
                                    "caption": {
                                        "type": "string",
                                        "example": "Nova caption editada! 📸 #atualizando",
                                        "maxLength": 2200
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "caption": {
                                            "type": "string",
                                            "example": "Nova caption editada! 📸 #atualizando"
                                        },
                                        "image_url": {
                                            "type": "string",
                                            "example": "/storage/posts/image_1.jpg"
                                        },
                                        "likes_count": {
                                            "type": "integer",
                                            "example": 42
                                        },
                                        "comments_count": {
                                            "type": "integer",
                                            "example": 8
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-05-22T10:30:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-05-22T11:45:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso negado (apenas o autor pode atualizar)"
                    },
                    "404": {
                        "description": "Post não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Deletar post",
                "description": "Remove um post existente. Apenas o autor do post pode deletá-lo. Requer autenticação.",
                "operationId": "d477c7dc08966e54b1a008f3cf9c5439",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID do post a deletar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Post deletado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso negado (apenas o autor pode deletar)"
                    },
                    "404": {
                        "description": "Post não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}